py-fastapi-seed/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
black
